# feat(fann): implement Adam/AdamW optimizers and SIMD acceleration