--- a/src/lib.rs
+++ b/src/lib.rs
@@ -41,6 +41,7 @@ pub mod connection;
 pub mod errors;
 pub mod integration;
 pub mod layer;
+pub mod memory_manager;
 pub mod network;
 pub mod neuron;
 pub mod training;