--- a/src/training/quickprop.rs
+++ b/src/training/quickprop.rs
@@ -1,5 +1,7 @@
 //! Quickprop training algorithm
 
+#![allow(clippy::needless_range_loop)]
+
 use super::*;
 use num_traits::Float;
 use std::collections::HashMap;