… Cargo.lock
… Cargo.toml
… examples/final_performance_demo.rs
… examples/test_adam.rs
… examples/test_optimizers_simple.rs
… src/training/rprop.rs
… src/training/test_all_algorithms.rs
… src/webgpu/autonomous_gpu_resource_manager.rs
… src/webgpu/device.rs
… src/webgpu/kernel_optimizer.rs
… src/webgpu/memory.rs
… src/webgpu/webgpu_backend.rs